requirements.txt
python-dotenv==1.2.1
langchain-google-genai==3.0.0
langchain==1.0.2
langgraph==1.0.1
fastapi
uvicorn